update main loop to new queue API
Calling `peek_event()` when the `things_to_do` queue is empty
trips a Boost assertion. This update checks to make sure there is
something in the queue before calling `peek_event()`.

Also, `eat_event()` calls `pop()` on the queue, which removes the item
from the queue. However, it doesn't update the internal state of the
item, in particular it doesn't un-mark the item as in a queue. Later on,
this causes the `reenqueue()` function to not re-add the item to the
queue, causing the event to never get triggered again. To fix this, we
update the state of the event after calling `eat_event()`.